Restructure XCP-D.
