+++ xcp_d/workflows/anatomical/__init__.py
@@ -1,0 +1,10 @@
+"""Workflow for anatomical postprocessing."""
+
+from xcp_d.workflows.anatomical import outputs, parcellation, surface, volume
+
+__all__ = [
+    "outputs",
+    "parcellation",
+    "surface",
+    "volume",
+]
