+++ xcp_d/workflows/bold/__init__.py
@@ -1,0 +1,19 @@
+"""Workflows for processing BOLD data."""
+
+from xcp_d.workflows.bold import (
+    cifti,
+    concatenation,
+    connectivity,
+    metrics,
+    nifti,
+    outputs,
+)
+
+__all__ = [
+    "cifti",
+    "concatenation",
+    "connectivity",
+    "metrics",
+    "nifti",
+    "outputs",
+]
