+++ xcp_d/workflows/bold/cifti.py
@@ -1,6 +1,6 @@
 # emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
 # vi: set ft=python sts=4 ts=4 sw=4 et:
-"""Workflows for post-processing the BOLD data."""
+"""Workflows for post-processing CIFTI-format BOLD data."""
 from nipype import logging
 from nipype.interfaces import utility as niu
 from nipype.pipeline import engine as pe
@@ -12,22 +12,22 @@ from xcp_d.interfaces.utils import ConvertTo32
 from xcp_d.utils.confounds import get_custom_confounds
 from xcp_d.utils.doc import fill_doc
 from xcp_d.utils.utils import _create_mem_gb
-from xcp_d.workflows.connectivity import init_functional_connectivity_nifti_wf
-from xcp_d.workflows.execsummary import init_execsummary_functional_plots_wf
-from xcp_d.workflows.outputs import init_postproc_derivatives_wf
-from xcp_d.workflows.plotting import init_qc_report_wf
-from xcp_d.workflows.postprocessing import (
+from xcp_d.workflows.bold.connectivity import init_functional_connectivity_cifti_wf
+from xcp_d.workflows.bold.metrics import init_alff_wf, init_reho_cifti_wf
+from xcp_d.workflows.bold.outputs import init_postproc_derivatives_wf
+from xcp_d.workflows.bold.postprocessing import (
     init_denoise_bold_wf,
     init_despike_wf,
     init_prepare_confounds_wf,
 )
-from xcp_d.workflows.restingstate import init_alff_wf, init_reho_nifti_wf
+from xcp_d.workflows.execsummary import init_execsummary_functional_plots_wf
+from xcp_d.workflows.plotting import init_qc_report_wf
 
 LOGGER = logging.getLogger("nipype.workflow")
 
 
 @fill_doc
-def init_postprocess_nifti_wf(
+def init_postprocess_cifti_wf(
     bold_file,
     head_radius,
     run_data,
@@ -35,9 +35,9 @@ def init_postprocess_nifti_wf(
     t2w_available,
     n_runs,
     exact_scans,
-    name="bold_postprocess_wf",
+    name="cifti_postprocess_wf",
 ):
-    """Organize the bold processing workflow.
+    """Organize the cifti processing workflow.
 
     Workflow Graph
         .. workflow::
@@ -49,25 +49,22 @@ def init_postprocess_nifti_wf(
             from xcp_d.tests.tests import mock_config
             from xcp_d import config
             from xcp_d.utils.bids import collect_data, collect_run_data
-            from xcp_d.workflows.bold import init_postprocess_nifti_wf
+            from xcp_d.workflows.bold.cifti import init_postprocess_cifti_wf
 
             with mock_config():
                 bold_file = str(
                     config.execution.fmri_dir / "sub-01" / "func" /
-                    (
-                        "sub-01_task-imagery_run-01_space-MNI152NLin2009cAsym_res-2_"
-                        "desc-preproc_bold.nii.gz"
-                    )
+                    "sub-01_task-imagery_run-01_space-fsLR_den-91k_bold.dtseries.nii"
                 )
 
                 run_data = collect_run_data(
                     layout=layout,
                     input_type="fmriprep",
                     bold_file=bold_file,
-                    cifti=False,
+                    cifti=True,
                 )
 
-                wf = init_postprocess_nifti_wf(
+                wf = init_postprocess_cifti_wf(
                     bold_file=bold_file,
                     head_radius=50.,
                     run_data=run_data,
@@ -75,13 +72,14 @@ def init_postprocess_nifti_wf(
                     t2w_available=True,
                     n_runs=1,
                     exact_scans=[],
-                    name="nifti_postprocess_wf",
+                    name="cifti_postprocess_wf",
                 )
 
     Parameters
     ----------
-    bold_file: :obj:`str`
-        bold file for post processing
+    bold_file
+    %(head_radius)s
+        This will already be estimated before this workflow.
     run_data : dict
     t1w_available
     t2w_available
@@ -90,28 +88,19 @@ def init_postprocess_nifti_wf(
         This is just used for the boilerplate, as this workflow only posprocesses one run.
     %(exact_scans)s
     %(name)s
-        Default is "nifti_postprocess_wf".
+        Default is "cifti_postprocess_wf".
 
     Inputs
     ------
     bold_file
-        BOLD series NIfTI file
+        CIFTI file
     %(boldref)s
-        Loaded in this workflow.
-    bold_mask
-        bold_mask from fmriprep
-        Loaded in this workflow.
     %(custom_confounds_file)s
-    %(template_to_anat_xfm)s
-        Fed from the subject workflow.
     t1w
         Preprocessed T1w image, warped to standard space.
         Fed from the subject workflow.
     t2w
         Preprocessed T2w image, warped to standard space.
-        Fed from the subject workflow.
-    anat_brainmask
-        T1w brain mask in standard space, used for transforms in the QC report workflow.
         Fed from the subject workflow.
     %(fmriprep_confounds_file)s
     fmriprep_confounds_json
@@ -122,17 +111,18 @@ def init_postprocess_nifti_wf(
     %(name_source)s
     preprocessed_bold : :obj:`str`
         The preprocessed BOLD file, after dummy scan removal.
-    %(filtered_motion)s
-    %(temporal_mask)s
     %(fmriprep_confounds_file)s
         After dummy scan removal.
+    %(filtered_motion)s
+    %(temporal_mask)s
     %(denoised_interpolated_bold)s
+    %(censored_denoised_bold)s
     %(smoothed_denoised_bold)s
     %(boldref)s
     bold_mask
+        This will not be defined.
     %(timeseries)s
     %(timeseries_ciftis)s
-        This will not be defined.
 
     References
     ----------
@@ -154,13 +144,9 @@ def init_postprocess_nifti_wf(
             fields=[
                 "bold_file",
                 "boldref",
-                "bold_mask",
                 "custom_confounds_file",
-                "template_to_anat_xfm",
                 "t1w",
                 "t2w",
-                "anat_native",
-                "anat_brainmask",
                 "fmriprep_confounds_file",
                 "fmriprep_confounds_json",
                 "dummy_scans",
@@ -168,6 +154,9 @@ def init_postprocess_nifti_wf(
                 "atlases",
                 "atlas_files",
                 "atlas_labels_files",
+                # for plotting, if the anatomical workflow was used
+                "lh_midthickness",
+                "rh_midthickness",
             ],
         ),
         name="inputnode",
@@ -175,7 +164,6 @@ def init_postprocess_nifti_wf(
 
     inputnode.inputs.bold_file = bold_file
     inputnode.inputs.boldref = run_data["boldref"]
-    inputnode.inputs.bold_mask = run_data["boldmask"]
     inputnode.inputs.fmriprep_confounds_file = run_data["confounds"]
     inputnode.inputs.fmriprep_confounds_json = run_data["confounds_json"]
     inputnode.inputs.dummy_scans = dummy_scans
@@ -188,6 +176,8 @@ def init_postprocess_nifti_wf(
         custom_confounds_folder,
         run_data["confounds"],
     )
+
+    workflow = Workflow(name=name)
 
     workflow.__desc__ = f"""\
 
@@ -209,10 +199,10 @@ the following post-processing was performed.
                 "censored_denoised_bold",
                 "smoothed_denoised_bold",
                 "boldref",
-                "bold_mask",
+                "bold_mask",  # will not be defined
                 # if parcellation is performed
                 "timeseries",
-                "timeseries_ciftis",  # will not be defined
+                "timeseries_ciftis",
             ],
         ),
         name="outputnode",
@@ -228,16 +218,11 @@ the following post-processing was performed.
     )
 
     workflow.connect([
-        (inputnode, outputnode, [("bold_file", "name_source")]),
-        (inputnode, downcast_data, [
-            ("bold_file", "bold_file"),
-            ("boldref", "boldref"),
-            ("bold_mask", "bold_mask"),
-        ]),
-        (downcast_data, outputnode, [
-            ("bold_mask", "bold_mask"),
+        (inputnode, outputnode, [
+            ("bold_file", "name_source"),
             ("boldref", "boldref"),
         ]),
+        (inputnode, downcast_data, [("bold_file", "bold_file")]),
     ])  # fmt:skip
 
     prepare_confounds_wf = init_prepare_confounds_wf(
@@ -253,7 +238,9 @@ the following post-processing was performed.
             ("fmriprep_confounds_file", "inputnode.fmriprep_confounds_file"),
             ("fmriprep_confounds_json", "inputnode.fmriprep_confounds_json"),
         ]),
-        (downcast_data, prepare_confounds_wf, [("bold_file", "inputnode.preprocessed_bold")]),
+        (downcast_data, prepare_confounds_wf, [
+            ("bold_file", "inputnode.preprocessed_bold"),
+        ]),
         (prepare_confounds_wf, outputnode, [
             ("outputnode.fmriprep_confounds_file", "fmriprep_confounds_file"),
             ("outputnode.preprocessed_bold", "preprocessed_bold"),
@@ -263,7 +250,6 @@ the following post-processing was performed.
     denoise_bold_wf = init_denoise_bold_wf(TR=TR, mem_gb=mem_gbx)
 
     workflow.connect([
-        (downcast_data, denoise_bold_wf, [("bold_mask", "inputnode.mask")]),
         (prepare_confounds_wf, denoise_bold_wf, [
             ("outputnode.temporal_mask", "inputnode.temporal_mask"),
             ("outputnode.confounds_file", "inputnode.confounds_file"),
@@ -297,7 +283,10 @@ the following post-processing was performed.
         alff_wf = init_alff_wf(name_source=bold_file, TR=TR, mem_gb=mem_gbx)
 
         workflow.connect([
-            (downcast_data, alff_wf, [("bold_mask", "inputnode.bold_mask")]),
+            (inputnode, alff_wf, [
+                ("lh_midthickness", "inputnode.lh_midthickness"),
+                ("rh_midthickness", "inputnode.rh_midthickness"),
+            ]),
             (prepare_confounds_wf, alff_wf, [
                 ("outputnode.temporal_mask", "inputnode.temporal_mask"),
             ]),
@@ -306,10 +295,13 @@ the following post-processing was performed.
             ]),
         ])  # fmt:skip
 
-    reho_wf = init_reho_nifti_wf(name_source=bold_file, mem_gb=mem_gbx)
+    reho_wf = init_reho_cifti_wf(name_source=bold_file, mem_gb=mem_gbx)
 
     workflow.connect([
-        (downcast_data, reho_wf, [("bold_mask", "inputnode.bold_mask")]),
+        (inputnode, reho_wf, [
+            ("lh_midthickness", "inputnode.lh_midthickness"),
+            ("rh_midthickness", "inputnode.rh_midthickness"),
+        ]),
         (denoise_bold_wf, reho_wf, [
             ("outputnode.censored_denoised_bold", "inputnode.denoised_bold"),
         ]),
@@ -322,14 +314,7 @@ the following post-processing was performed.
     )
 
     workflow.connect([
-        (inputnode, qc_report_wf, [
-            ("bold_file", "inputnode.name_source"),
-            ("boldref", "inputnode.boldref"),
-            ("bold_mask", "inputnode.bold_mask"),
-            ("anat_native", "inputnode.anat"),
-            ("anat_brainmask", "inputnode.anat_brainmask"),
-            ("template_to_anat_xfm", "inputnode.template_to_anat_xfm"),
-        ]),
+        (inputnode, qc_report_wf, [("bold_file", "inputnode.name_source")]),
         (prepare_confounds_wf, qc_report_wf, [
             ("outputnode.preprocessed_bold", "inputnode.preprocessed_bold"),
             ("outputnode.dummy_scans", "inputnode.dummy_scans"),
@@ -355,6 +340,11 @@ the following post-processing was performed.
             ("fmriprep_confounds_file", "inputnode.fmriprep_confounds_file"),
             ("atlas_files", "inputnode.atlas_files"),
         ]),
+        (denoise_bold_wf, postproc_derivatives_wf, [
+            ("outputnode.denoised_bold", "inputnode.denoised_bold"),
+            ("outputnode.smoothed_denoised_bold", "inputnode.smoothed_denoised_bold"),
+        ]),
+        (qc_report_wf, postproc_derivatives_wf, [("outputnode.qc_file", "inputnode.qc_file")]),
         (prepare_confounds_wf, postproc_derivatives_wf, [
             ("outputnode.confounds_file", "inputnode.confounds_file"),
             ("outputnode.confounds_metadata", "inputnode.confounds_metadata"),
@@ -363,11 +353,6 @@ the following post-processing was performed.
             ("outputnode.temporal_mask", "inputnode.temporal_mask"),
             ("outputnode.temporal_mask_metadata", "inputnode.temporal_mask_metadata"),
         ]),
-        (denoise_bold_wf, postproc_derivatives_wf, [
-            ("outputnode.denoised_bold", "inputnode.denoised_bold"),
-            ("outputnode.smoothed_denoised_bold", "inputnode.smoothed_denoised_bold"),
-        ]),
-        (qc_report_wf, postproc_derivatives_wf, [("outputnode.qc_file", "inputnode.qc_file")]),
         (reho_wf, postproc_derivatives_wf, [("outputnode.reho", "inputnode.reho")]),
         (postproc_derivatives_wf, outputnode, [
             ("outputnode.filtered_motion", "filtered_motion"),
@@ -375,6 +360,7 @@ the following post-processing was performed.
             ("outputnode.denoised_bold", "denoised_bold"),
             ("outputnode.smoothed_denoised_bold", "smoothed_denoised_bold"),
             ("outputnode.timeseries", "timeseries"),
+            ("outputnode.timeseries_ciftis", "timeseries_ciftis"),
         ]),
     ])  # fmt:skip
 
@@ -387,7 +373,10 @@ the following post-processing was performed.
         ])  # fmt:skip
 
     if atlases:
-        connectivity_wf = init_functional_connectivity_nifti_wf(mem_gb=mem_gbx)
+        connectivity_wf = init_functional_connectivity_cifti_wf(
+            mem_gb=mem_gbx,
+            exact_scans=exact_scans,
+        )
 
         workflow.connect([
             (inputnode, connectivity_wf, [
@@ -395,8 +384,9 @@ the following post-processing was performed.
                 ("atlases", "inputnode.atlases"),
                 ("atlas_files", "inputnode.atlas_files"),
                 ("atlas_labels_files", "inputnode.atlas_labels_files"),
+                ("lh_midthickness", "inputnode.lh_midthickness"),
+                ("rh_midthickness", "inputnode.rh_midthickness"),
             ]),
-            (downcast_data, connectivity_wf, [("bold_mask", "inputnode.bold_mask")]),
             (prepare_confounds_wf, connectivity_wf, [
                 ("outputnode.temporal_mask", "inputnode.temporal_mask"),
             ]),
@@ -405,6 +395,10 @@ the following post-processing was performed.
             ]),
             (reho_wf, connectivity_wf, [("outputnode.reho", "inputnode.reho")]),
             (connectivity_wf, postproc_derivatives_wf, [
+                ("outputnode.coverage_ciftis", "inputnode.coverage_ciftis"),
+                ("outputnode.timeseries_ciftis", "inputnode.timeseries_ciftis"),
+                ("outputnode.correlation_ciftis", "inputnode.correlation_ciftis"),
+                ("outputnode.correlation_ciftis_exact", "inputnode.correlation_ciftis_exact"),
                 ("outputnode.coverage", "inputnode.coverage"),
                 ("outputnode.timeseries", "inputnode.timeseries"),
                 ("outputnode.correlations", "inputnode.correlations"),
@@ -424,15 +418,15 @@ the following post-processing was performed.
     if config.workflow.abcc_qc:
         # executive summary workflow
         execsummary_functional_plots_wf = init_execsummary_functional_plots_wf(
-            preproc_nifti=bold_file,
+            preproc_nifti=run_data["nifti_file"],
             t1w_available=t1w_available,
             t2w_available=t2w_available,
             mem_gb=mem_gbx,
         )
 
         workflow.connect([
-            # Use inputnode for executive summary instead of downcast_data
-            # because T1w is used as name source.
+            # Use inputnode for executive summary instead of downcast_data because T1w is name
+            # source.
             (inputnode, execsummary_functional_plots_wf, [
                 ("boldref", "inputnode.boldref"),
                 ("t1w", "inputnode.t1w"),
